+++ src/game_api/script/script_impl.cpp
@@ -552,7 +552,7 @@ ScriptImpl::ScriptImpl(std::string script, std::string file, SoundManager* sound
     lua["testflag"] = lua["test_flag"];
 
     /// Gets the resolution (width and height) of the screen
-    lua["get_window_size"] = [this]() -> std::tuple<int, int> { return std::make_tuple(ImGui::GetWindowWidth(), ImGui::GetWindowHeight()); };
+    lua["get_window_size"] = [this]() -> std::tuple<int, int> { return {(int)ImGui::GetWindowWidth(), (int)ImGui::GetWindowHeight()}; };
 
     /// Steal input from a Player or HH.
     lua["steal_input"] = [this](int uid)
@@ -1154,48 +1154,6 @@ std::string ScriptImpl::script_id()
 {
     std::string newid = sanitize(meta.author) + "/" + sanitize(meta.name);
     return newid;
-}
-
-template <class... Args>
-bool ScriptImpl::handle_function(sol::function func, Args&&... args)
-{
-    return handle_function_with_return<std::monostate>(std::move(func), std::forward<Args>(args)...) != std::nullopt;
-}
-template <class Ret, class... Args>
-std::optional<Ret> ScriptImpl::handle_function_with_return(sol::function func, Args&&... args)
-{
-    auto lua_result = func(std::forward<Args>(args)...);
-    if (!lua_result.valid())
-    {
-        sol::error e = lua_result;
-        result = e.what();
-#ifdef SPEL2_EXTRA_ANNOYING_SCRIPT_ERRORS
-        messages.push_back({result, std::chrono::system_clock::now(), ImVec4(1.0f, 0.2f, 0.2f, 1.0f)});
-        DEBUG("{}", result);
-        if (messages.size() > 20)
-            messages.pop_front();
-#endif
-        return std::nullopt;
-    }
-    if constexpr (std::is_same_v<Ret, std::monostate>)
-    {
-        return std::optional{std::monostate{}};
-    }
-    else
-    {
-        try
-        {
-            auto return_type = lua_result.get_type();
-            return return_type == sol::type::none || return_type == sol::type::nil
-                       ? std::optional<Ret>{}
-                       : std::optional{static_cast<Ret>(lua_result)};
-        }
-        catch (...)
-        {
-            result = "Unexpected return type from function.";
-        }
-    }
-    return std::nullopt;
 }
 
 void ScriptImpl::render_options()
