+++ src/spel2_dll/spel2.cpp
@@ -3,7 +3,6 @@
 #include "file_api.hpp"
 #include "rpc.hpp"
 #include "script.hpp"
-#include "script/usertypes/save_context.hpp"
 #include "sound_manager.hpp"
 #include "state.hpp"
 #include "window_api.hpp"
